backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
transformers
torch
Pillow
pymongo
timm
